requirements.txt
alabaster==0.7.13
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.12.1
billiard==3.6.4.0
celery==5.2.7
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
commonmark==0.9.1
constantly==23.10.4
cryptography==42.0.5
daphne==4.0.0
distlib==0.3.6
Django==4.0.8
django-celery-beat==2.3.0
django-celery-results==2.3.1
django-ninja==0.19.1
django-redis==5.2.0
django-timezone-field==6.1.0
docutils==0.18.1
et-xmlfile==1.1.0
filelock==3.12.2
hyperlink==21.0.0
idna==3.4
imagesize==1.4.1
incremental==22.10.0
Jinja2==3.1.2
kombu==5.3.7
MarkupSafe==2.1.3
openpyxl==3.0.10
packaging==23.1
pbr==5.11.1
platformdirs==3.8.0
prompt-toolkit==3.0.43
psutil==5.9.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==1.9.2
Pygments==2.16.1
PyMySQL==1.1.0
pyOpenSSL==24.1.0
pypinyin==0.46.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.11
PyQt5-sip==12.13.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
recommonmark==0.7.1
redis==5.0.4
requests==2.28.1
service-identity==24.1.0
simplejwt==2.0.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.5.0
stevedore==5.1.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
ua-parser==0.18.0
urllib3==1.26.18
user-agents==2.2.0
vine==5.1.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.13
zope.interface==6.3
